Addressing Compilation Issue (interface level)
This change provides a check for a compilation failure at the
interface level.

Previously if students' code was not able to be compiled with the unit
testing test suite in M3, they would get a 0 for the assignment instead of
getting graded on the application-level (python tests).

Disclaimer: This solution only really applies for M3. It assumes that
1) Interface level testing occurs at some point
2) The unit-testing framework used was CUnit

Some other solutions to this problem include:
1. Backward-compatible new main.py which accepts a new optional arg to specify
the existence of unit tests

2. Modifications to config.json as well as a new python "interface" file that
would need to be overridden by the AAT developer. Could also include the backward
compatability to main.py.